Circular dependency fix?
