+++ src/shared/sam/cli/samCliProcessInvokerContext.ts
@@ -1,0 +1,20 @@
+/*!
+ * Copyright 2020 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+import { extensionSettingsPrefix } from '../../constants'
+import { DefaultSettingsConfiguration } from '../../settingsConfiguration'
+import { DefaultSamCliConfiguration, SamCliConfiguration } from './samCliConfiguration'
+import { DefaultSamCliLocationProvider } from './samCliLocator'
+
+export interface SamCliProcessInvokerContext {
+    cliConfig: SamCliConfiguration
+}
+
+export class DefaultSamCliProcessInvokerContext implements SamCliProcessInvokerContext {
+    public cliConfig: SamCliConfiguration = new DefaultSamCliConfiguration(
+        new DefaultSettingsConfiguration(extensionSettingsPrefix),
+        new DefaultSamCliLocationProvider()
+    )
+}
